+++ src/vws/query.py
@@ -154,6 +154,7 @@ class CloudRecoService:
             status_code=requests_response.status_code,
             headers=dict(requests_response.headers),
             request_body=requests_response.request.body,
+            raw=requests_response.raw,
         )
 
         if response.status_code == HTTPStatus.REQUEST_ENTITY_TOO_LARGE: