+++ src/vws/vws.py
@@ -125,6 +125,7 @@ def _target_api_request(
         status_code=requests_response.status_code,
         headers=dict(requests_response.headers),
         request_body=requests_response.request.body,
+        raw=requests_response.raw,
     )